Part 7 add the functionality to be able to upload photos to our API add 2 additonal methods in here as well. One method to remove a photo from the Photos collection for a product and also a method that will set a photo to be the main photo. Add the following additional methods
We want to be able to save the photos to disk – in order to do this we will create a PhotoService and corresponding interface here so that we do not have the saving to disk logic inside our controllers. Add a new Interface called IPhotoService.cs in the Interfaces folder inside the Core project

You will get an error here as the IFormFile is not recognised. In order to make use of this class we need to add an additonal nuget package to the Core project so go ahead and use Nuget package manager to add the Microsoft.AspNetCore.Http.Features   --version 3.1.3 package or run the following command inside the Core project folder

Next lets add the implementation class for this. Add a new class called PhotoService.cs in the Services folder in the Infrastructure project

Next up, lets create a DTO that we will use as our parameter in the AddProductPhoto method. In the Dtos folder create a new class called ProductPhotoDto.cs

We also need to add the new PhotoService to our Startup configuration so that we can inject it where we need to. Add the following to the ApplicationServiceExtensions.cs class in the Extensions folder

We also want to add some validation here to make sure that we only receive files with that are either Jpgs or Pngs and since we don’t want users to be able to upload gigantic files we also want to prevent our server accepting them There are no default validators for this so we need to make our own
So lets create 2 new classes for some custom validation attributes – one for each type of validation so each one just has a single responsibility. First create a new class in the Helpers folder in the API called AllowedExtensionsAttribute.cs

Then in order to use these we can just update the ProductPhotoDto.cs

Now our API server will only accept files that are less than 2Mb in size and have a file extension of either .jpg, .png or .jpeg.

With this in place we can now add the functionality we need to the ProductsController.cs. First of all lets add the AddProductPhoto method and inject our new IPhotoService into the controller

so we can upload images. Lets now work on removing an image. We will not let the main photo be deleted but all the others will be fair game for deletion. Lets now add the DeleteProductPhoto method

One more method to go, the SetMainPhoto method

We should now be able to set any photo that is not already the main photo for a product to be so