Part 8  In this part we will add the ability to upload product images, but first – since we now have the ability to store multiple images for a product – lets add a gallery into the product details section Install ngx-gallery to the app  Run in the client folder:
npm install @kolkov/ngx-gallery --force

Then lets import this into the Shared module and export it for use

We also need to update our product.ts as we now are getting a photo collection along with the products from the API. Update the product.ts class for the IProduct and the IPhoto

Now we can update the product-details.component.ts to make use of the gallery package that we have installed

We should now see that we have our gallery in place – admittedly things do not look too much different as we only have a single image for each product.

But it’s a start. In the next part we will look at how we can upload images to the API